P10.c

+++ P10.c
@@ -1,20 +1,29 @@
 #include <stdio.h>
 int main() {
+    
     int math, eng, sci, hindi, pe, totalMarks;
     float percentage;
+
     printf("ENTER MARKS FOR MATHS: ");
     scanf("%d", &math);
+
     printf("ENTER MARKS FOR ENGLISH: ");
     scanf("%d", &eng);
+
     printf("ENTER MARKS FOR SCIENCE: ");
     scanf("%d", &sci);
+
     printf("ENTER MARKS FOR HINDI: ");
     scanf("%d", &hindi);
+
     printf("ENTER MARKS FOR P.E: ");
     scanf("%d", &pe);
+
     totalMarks = math + eng + sci + hindi + pe;
-    percentage = (float)totalMarks / 5.0;
+    percentage = totalMarks / 5.0;
+    
     printf("Percentage = %.2f\n", percentage);
+    
     if(percentage >= 90)
     {
         printf("Grade A\n");
@@ -31,4 +40,5 @@ int main() {
     {
         printf("Grade D\n");
     }
+
 }
